New subtyping rules And-Typ and Or-Typ to simplify type members
The new rules are:
 Γ ⊢ (L: T₁..T₂) ∧ (L: T₃..T₄) <: (L: (T₁ ∨ T₃)..(T₂ ∧ T₄)) (And-Typ)
 Γ ⊢ (L: T₁..T₂) ∨ (L: T₃..T₄) <: (L: (T₁ ∧ T₃)..(T₂ ∨ T₄)) (Or-Typ)

The first one is needed to define a subtype-preserving translation of type members as
they exist in Scala.

The second one isn't actually needed because in Scala, we never directly select
members on a union (we always widen a union to an intersection of the common
parent classes of its operands first), so it is present here just for symmetry.